Issue Management Workflows
There are a lot of issues pouring in, mostly from first time open-source contributors.

These workflows should help clean up all the invalid and spam issues.

Note: This is more of a suggestion than anything else. There are a lot of invalid/spam issues and it may be hard to cut through the noise and find legitmate issues that nee fixing. Semi-dependant on #342 and #300.