Remove conversion of keycodes in playstate
It seems these keycodes are already converted? I'm not sure, but this fixes a bug where you can't bind some keys, for example try rebinding up arrow to right and vice versa, it will not work as it's converting it to wrong keys, for example up is converted to the quote (') key.