Update Controls.hx
Changed Control.RESET from [Y] to [6]. 

Why? Because that means instead of pressing Y to reset the song, you instead press the select button, which prevents accidental resets if you are playing on a non-standard xbox controller. 

I brought up this issue back in 2022 and it never got solved so I just decided I'd make the pull request myself.