Remove `hmm` dependency to fix HTML5 target
Remove the `hmm` haxelib dependency from the build.

Linking to `hmm` caused `utest` to be transitively linked, which
eventually caused OpenFL to act strangely and instatiate the
application twice on the HTML5 target.

`hmm` was only used for `HaxelibVersions.getLibraryVersions` macro
call. Instead, manually parse the `hmm.json` ourselves to avoid
the dependency. This fixes the HTML5 target.

`hmm` is still used for package management, but no longer linked
in to the build itself.