Remove DynamicTools; fix pause menu on HTML5
Calls intended for `ArrayTools.clone` were being routed to
`DynamicTools.clone` due to the order of `using` statements in
`imports.hx`. This caused the pause menu to break due to arrays
becoming fubar (missing length property).

Using `DynamicTools` is a little dangerous, so remove it in favor
of calling `Reflect.copy` directly.