Fix bad initial save data when no save data present
If no save data was present, the game would check for legacy save
data in the ninjamuffin99 path to migrate over. This code was
incorrectly checking `FlxSave.data == null` to determine if
legacy data was present.

This caused an empty legacy save being migrated over, resulting
in a `volume` of 0 for any new player of the game.