save stuff
Note: The previous Enums DON'T get saved as Enums due to the Serializer Bug. Instead they get saved as Abstract Enums. The previous Enums were made deprecated as a form of backwards compatibility, however due to the Serializer Bug it is impossible to save current Enum Abstracts and previous Enums, and thus they are the only fields that get reset. To prevent this, a change to haxe's Serializer would have to be made.