last adjustment before I'm brutally murdered by the Funkin'™ GitHub mods
this changes it so the editor is more responsible for updating the hold notes, to make it similar to how it handles tap notes.
I still kept the position update on overrideStepTime to match the tap note behavior.
Also now the hold note's direction is updated along with the position

And finally replaced an bigass if-else expression with an utility function that's also used by the tap note sprite code.